- Fix sign of comparison warning in shared ports setup.

+++ services/outside_network.c
@@ -4128,14 +4128,14 @@ static int shared_ports_alloc_ifs(struct shared_ports* shp, char** ifs,
 		int i;
 		for(i=0; i<num_ifs; i++) {
 			if(str_is_ip6(ifs[i]) && do_ip6 &&
-				done_6 < shp->num_ip6) {
+				(int)done_6 < shp->num_ip6) {
 				if(!shared_ports_setup_if(&shp->ip6_ifs[done_6],
 					ifs[i], availports, numavailports))
 					return 0;
 				done_6++;
 			}
 			if(!str_is_ip6(ifs[i]) && do_ip4 &&
-				done_4 < shp->num_ip4) {
+				(int)done_4 < shp->num_ip4) {
 				if(!shared_ports_setup_if(&shp->ip4_ifs[done_4],
 					ifs[i], availports, numavailports))
 					return 0;
